electron/extraResources/edith_backend/requirements.txt
# Core — pinned to exact tested versions
python-dotenv==1.2.1
google-genai==1.64.0
tqdm>=4.66
watchdog>=4.0
reportlab>=4.0
cryptography>=42.0
pypdf==6.7.0
PyMuPDF==1.27.1
chromadb==1.5.0
sentence-transformers==5.2.2
python-docx>=1.1.2
pdf2image>=1.17.0
pytesseract>=0.3.10
fastapi==0.115.9
uvicorn==0.40.0
pydantic==2.12.5
python-multipart>=0.0.9
httpx==0.28.1
starlette==0.45.3

# Security
PyJWT>=2.8.0

# Model integrations
requests>=2.31.0
openai==2.21.0
onnxruntime==1.24.1

# Training / eval
scikit-learn>=1.3.0
